requirements.txt
dash
pandas
plotly
numpy
requests
pyyaml
schedule
gpsd-py3
csv-logger
pytz
